Fixed startup logic and cleaned MainActivity, gradle build files.
